Added download_artist_albums
Added a function to be called in client if
artist_id_str is not None;
download_artist_albums(artist_id_str) that
uses the get_artist_album function to get the
artist's album id's and then downloads each one
with the download_album function.

Also modified get_artist_album function to return
a list of album ids instead of a list of tuples
of album ids and names.