Added argument parsing in search function
Added a block for parsing arguments passed with options
-l -limit
-t -type
in the search() function. Arguments are passed inside search_term.
They work as follows:

* -l -limit <number> : sets the limit of results to that number
  and raises a ValueError if that number exceeds 50. Default is 10.

* -t -type <album/track/playlist> : sets the type that is requested
  from the API about the search query. Raises a ValueError if an arguments
  passed is different than track, album, playlist. Default is all three.

Example:
Enter search or URL: <query> -l 30 -t track album

This will result with 30 tracks and 30 albums associated with query.
Options can be passed in any order but the query must be first.